copy_strided seems to build

+++ src/kernels-glsl/common/broadcast_dims.glsl
@@ -6,7 +6,7 @@
 ///
 ///////////////////////////////////////////////////////////////////////////////
 typedef struct Shape {
-    ulong s[DIMS];
+    uint s[DIMS];
 } Shape;
 
 
@@ -46,7 +46,7 @@ inline Shape get_pos_broadcast(Shape limits)
     r.s[2] = get_global_id(1) / limits.s[3];      
     r.s[3] = get_global_id(1) % limits.s[3];      
     r.s[4] = get_global_id(0) / (limits.s[5]*limits.s[6]);      
-    ulong s56 = get_global_id(0) % (limits.s[5]*limits.s[6]);
+    uint s56 = get_global_id(0) % (limits.s[5]*limits.s[6]);
     r.s[5] = s56 / limits.s[6];      
     r.s[6] = s56 % limits.s[6];      
 #elif DIMS == 8
@@ -54,12 +54,12 @@ inline Shape get_pos_broadcast(Shape limits)
     r.s[1] = get_global_id(2) % limits.s[1];      
 
     r.s[2] = get_global_id(1) / (limits.s[3]*limits.s[4]);      
-    ulong s34 = get_global_id(1) % (limits.s[3]*limits.s[4]);
+    uint s34 = get_global_id(1) % (limits.s[3]*limits.s[4]);
     r.s[3] = s34 / limits.s[4];      
     r.s[4] = s34 % limits.s[4];      
 
     r.s[5] = get_global_id(0) / (limits.s[6]*limits.s[7]);      
-    ulong s67 = get_global_id(0) % (limits.s[6]*limits.s[7]);
+    uint s67 = get_global_id(0) % (limits.s[6]*limits.s[7]);
     r.s[6] = s67 / limits.s[7];      
     r.s[7] = s67 % limits.s[7];      
 #else
